Rename xml => xml_parser
Because PyQt5 uses xml.elementTree (which is overshadowed because
it uses a generic dump top level name)
